refactor(choose-plan-page): Refactor IPaymentPageInteractor interface and usePaymentPageInteractor function
-  Removed IPaymentPageInteractor interface and integrated its structure directly into usePaymentPageInteractor function for better encapsulation and organization
